Add packages for bobby, command, exceptions, parser, storage, tasklist and tasks.
Previously all the files were stored under main/java. However, this was very messy as more classes are added.

Add packages to group related classes together to make the code neater.
