+++ src/main/java/bobby/command/Command.java
@@ -1,3 +1,5 @@
+package bobby.command;
+
 public enum Command {
     BYE, LIST, MARK, UNMARK, DELETE, TODO, DEADLINE, EVENT, UNKNOWN, FIND;
