+++ src/main/java/bobby/exceptions/BobbyException.java
@@ -1,0 +1,8 @@
+package bobby.exceptions;
+
+public class BobbyException extends Exception {
+    public BobbyException(String message) {
+        super(message);
+    }
+}
+
