+++ src/main/java/bobby/exceptions/EmptyEventException.java
@@ -1,0 +1,7 @@
+package bobby.exceptions;
+
+public class EmptyEventException extends BobbyException {
+    public EmptyEventException() {
+        super("The description, start time, or end time of an event cannot be empty.");
+    }
+}
