+++ src/main/java/bobby/exceptions/EmptyTodoException.java
@@ -1,0 +1,7 @@
+package bobby.exceptions;
+
+public class EmptyTodoException extends BobbyException {
+    public EmptyTodoException() {
+        super("The description of todo cannot be empty!");
+    }
+}
