+++ src/main/java/bobby/exceptions/InvalidDateException.java
@@ -1,0 +1,7 @@
+package bobby.exceptions;
+
+public class InvalidDateException extends BobbyException {
+    public InvalidDateException() {
+        super("Invalid date format. Please use yyyy-MM-dd.");
+    }
+}
