+++ src/main/java/bobby/exceptions/InvalidInputException.java
@@ -1,0 +1,7 @@
+package bobby.exceptions;
+
+public class InvalidInputException extends BobbyException {
+    public InvalidInputException() {
+        super("I am sorry, but I do not know what that means!");
+    }
+}
