+++ src/main/java/bobby/exceptions/InvalidTaskNumberException.java
@@ -1,0 +1,7 @@
+package bobby.exceptions;
+
+public class InvalidTaskNumberException extends BobbyException {
+    public InvalidTaskNumberException() {
+        super("The task number provided is invalid.");
+    }
+}
