+++ src/main/java/bobby/tasks/Deadline.java
@@ -1,3 +1,5 @@
+package bobby.tasks;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
