+++ src/main/java/bobby/tasks/Task.java
@@ -1,3 +1,5 @@
+package bobby.tasks;
+
 public class Task {
     private String description;
     private boolean isDone;
